Moved mock framework into commons
