+++ galasa-parent/dev.galasa.framework.api.cps/src/test/java/dev/galasa/framework/api/cps/internal/CpsServletTest.java
@@ -8,11 +8,10 @@ package dev.galasa.framework.api.cps.internal;
 import com.google.gson.*;
 
 import dev.galasa.framework.api.cps.internal.mocks.MockCpsServlet;
-import dev.galasa.framework.api.cps.internal.mocks.MockFramework;
-import dev.galasa.framework.api.cps.internal.mocks.MockIConfigurationPropertyStoreService;
-
+import dev.galasa.framework.api.common.mocks.MockFramework;
 import dev.galasa.framework.api.common.mocks.MockHttpServletRequest;
 import dev.galasa.framework.api.common.mocks.MockHttpServletResponse;
+import dev.galasa.framework.api.common.mocks.MockIConfigurationPropertyStoreService;
 import dev.galasa.framework.api.common.mocks.MockServletOutputStream;
 import dev.galasa.framework.spi.IConfigurationPropertyStoreService;
 import dev.galasa.framework.spi.IFramework;
