--- galasa-parent/dev.galasa.framework.api.cps/src/test/java/dev/galasa/framework/api/cps/internal/mocks/IServletUnderTest.java
@@ -1,15 +1,0 @@
-/*
- * Copyright contributors to the Galasa project
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package dev.galasa.framework.api.cps.internal.mocks;
-
-import dev.galasa.framework.IFileSystem;
-import dev.galasa.framework.spi.IFramework;
-
-public interface IServletUnderTest {
-    void setFramework(IFramework framework);
-
-    void setFileSystem(IFileSystem fileSystem);
-}
