+++ galasa-parent/dev.galasa.framework.api.cps/src/test/java/dev/galasa/framework/api/cps/internal/mocks/MockCpsServlet.java
@@ -6,6 +6,7 @@
 package dev.galasa.framework.api.cps.internal.mocks;
 
 import dev.galasa.framework.IFileSystem;
+import dev.galasa.framework.api.common.mocks.IServletUnderTest;
 import dev.galasa.framework.api.cps.internal.CpsServlet;
 import dev.galasa.framework.spi.IFramework;
