+++ galasa-parent/dev.galasa.framework.api.ras/src/test/java/dev/galasa/framework/api/ras/internal/mocks/MockRasServletEnvironment.java
@@ -7,6 +7,7 @@ package dev.galasa.framework.api.ras.internal.mocks;
 
 import java.util.List;
 
+import dev.galasa.framework.api.common.mocks.IServletUnderTest;
 import dev.galasa.framework.api.common.mocks.MockHttpServletRequest;
 import dev.galasa.framework.api.ras.internal.RasServlet;
 import dev.galasa.framework.mocks.MockFileSystem;
